chore: creating searchbar and contributors components
